+++ odradek-app/src/main/java/sh/adelessfox/odradek/app/ui/component/graph/GraphToolPanelEvent.java
@@ -1,0 +1,8 @@
+package sh.adelessfox.odradek.app.ui.component.graph;
+
+import sh.adelessfox.odradek.event.Event;
+
+public sealed interface GraphToolPanelEvent extends Event {
+    record UpdateFilter(String query, boolean matchCase, boolean matchWholeWord) implements GraphToolPanelEvent {
+    }
+}
