--- odradek-app/src/main/java/sh/adelessfox/odradek/app/ui/component/graph/GraphViewEvent.java
@@ -1,8 +1,0 @@
-package sh.adelessfox.odradek.app.ui.component.graph;
-
-import sh.adelessfox.odradek.event.Event;
-
-public sealed interface GraphViewEvent extends Event {
-    record UpdateFilter(String query, boolean matchCase, boolean matchWholeWord) implements GraphViewEvent {
-    }
-}
