+++ odradek-ui/src/main/resources/themes/FlatLaf.properties
@@ -4,7 +4,6 @@ FlatLaf.experimental.tree.widePathForLocation = true
 ToolTipUI = sh.adelessfox.odradek.ui.components.laf.FlatOutlineToolTipUI
 SplitButtonUI = sh.adelessfox.odradek.ui.components.laf.FlatSplitButtonUI
 LabeledSeparatorUI = sh.adelessfox.odradek.ui.components.laf.FlatLabeledSeparatorUI
-SplitterUI = sh.adelessfox.odradek.ui.components.laf.SplitterUI
 
 TitlePane.useWindowDecorations = true
 TitlePane.unifiedBackground = true
